@@ -3,7 +3,7 @@ description: A new Flutter project.
 version: 2.0.0+1
 
 environment:
-  sdk: '>=3.1.5 <4.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -23,13 +23,13 @@ dependencies:
   # For Model
   freezed_annotation: ^2.4.1
   go_router: ^12.1.1
-  shared_preferences: ^2.2.2
+  shared_preferences: ^2.3.3
   get_it: ^7.6.4
   dartz: ^0.10.1
   pretty_dio_logger: ^1.3.1
   internet_connection_checker: ^1.0.0+1
   easy_localization: ^3.0.3
-  device_preview: ^1.1.0
+  device_preview: ^1.2.0
   package_info_plus: ^8.0.0
   injectable: ^2.3.2
   flutter_dotenv: ^5.1.0
@@ -37,28 +37,26 @@ dependencies:
   retrofit: ^4.0.3
   bot_toast: ^4.1.3
   flutter_spinkit: ^5.2.0
-  qr_code_scanner: ^1.0.1
+  mobile_scanner: ^6.0.0
   cached_network_image: ^3.3.1
   crypto: ^3.0.3
-  flutter_inappwebview: ^6.0.0
+  flutter_inappwebview: ^6.1.5
 
-  file_picker: ^5.3.3
-  file_picker_pro: ^1.5.0
-  image_picker: ^1.0.7
+  file_picker: ^8.1.6
+  image_picker: ^1.1.2
   path: ^1.9.0
-  flutter_slidable: ^3.0.1
+  flutter_slidable: ^3.1.1
   font_awesome_flutter: ^10.7.0
   connectivity_plus: ^6.0.3
   flutter_screenutil: ^5.9.0
-  open_file_plus: ^3.4.1+1
+  open_filex: ^4.7.0
   timeago: ^3.6.1
-  path_provider: ^2.1.2
-  external_path: ^1.0.3
+  path_provider: ^2.1.5
   permission_handler: ^11.3.1
   hive: ^2.2.3
   hive_flutter: ^1.1.0
   html: ^0.15.4
-  flutter_secure_storage: ^9.0.0
+  flutter_secure_storage: ^9.2.4
 
 dev_dependencies:
   flutter_test:
@@ -95,6 +93,14 @@ flutter_gen:
     outputs:
       class_name: AppColors
 
+dependency_overrides:
+  win32: ^5.10.0
+  flutter_plugin_android_lifecycle: ^2.0.22
+  image_picker_android: ^0.8.12
+  path_provider_android: ^2.2.12
+  shared_preferences_android: ^2.3.3
+  permission_handler_android: ^12.0.12
+  sqflite: ^2.4.1
 
 flutter:
   uses-material-design: true
