@@ -1,6 +1,5 @@
 import 'dart:convert';
 import 'dart:developer';
-import 'dart:io';
 
 import 'package:axon_ivy/features/profile/domain/entities/qr/qr_model.dart';
 import 'package:axon_ivy/shared/widgets/back_button_widget.dart';
@@ -9,7 +8,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:go_router/go_router.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:qr_code_scanner/qr_code_scanner.dart';
+import 'package:mobile_scanner/mobile_scanner.dart';
 
 class QRParentPage extends StatefulWidget {
   const QRParentPage({super.key});
@@ -19,29 +18,25 @@ class QRParentPage extends StatefulWidget {
 }
 
 class _QRParentPageState extends State<QRParentPage> {
-  Barcode? result;
-  QRViewController? controller;
-  final GlobalKey qrKey = GlobalKey(debugLabel: 'QR');
-  late BuildContext qrViewContext;
+  String? result;
+  late final MobileScannerController controller;
+
   @override
-  void reassemble() {
-    super.reassemble();
-    if (Platform.isAndroid) {
-      controller!.pauseCamera();
-    }
-    controller!.resumeCamera();
+  void initState() {
+    super.initState();
+    controller = MobileScannerController();
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.background,
+        backgroundColor: Theme.of(context).colorScheme.surface,
         leadingWidth: 100.w,
         title: Text(
           "qr.title".tr(),
           style: GoogleFonts.inter(
-              color: Theme.of(context).colorScheme.surface,
+              color: Theme.of(context).colorScheme.onSurface,
               fontSize: 17.sp,
               fontWeight: FontWeight.w500),
         ),
@@ -52,58 +47,34 @@ class _QRParentPageState extends State<QRParentPage> {
   }
 
   Widget _buildQrView(BuildContext context) {
-    var scanArea = (MediaQuery.of(context).size.width < 400.w ||
-            MediaQuery.of(context).size.height < 400.h)
-        ? 150.0.h
-        : 300.0.h;
-    return QRView(
-      key: qrKey,
-      onQRViewCreated: _onQRViewCreated,
-      overlay: QrScannerOverlayShape(
-          borderColor: Colors.red,
-          borderRadius: 10.r,
-          borderLength: 30.r,
-          borderWidth: 10.r,
-          cutOutSize: scanArea),
-      onPermissionSet: (ctrl, p) => _onPermissionSet(context, ctrl, p),
-    );
-  }
+    return MobileScanner(
+      controller: controller,
+      onDetect: (BarcodeCapture capture) {
+        if (result != null) return;
+        final barcode = capture.barcodes.firstOrNull;
+        if (barcode == null || barcode.rawValue == null) return;
 
-  void _onQRViewCreated(QRViewController controller) {
-    setState(() {
-      this.controller = controller;
-    });
-    controller.scannedDataStream.listen((scanData) {
-      if (result == null) {
         setState(() {
-          result ??= scanData;
+          result = barcode.rawValue;
         });
-        debugPrint("${result?.code}");
-        if (result?.code != null) {
-          if (result!.code!.contains("loginUrl") &&
-              result!.code!.contains("username")) {
-            String jsonData = result!.code!;
+
+        debugPrint("$result");
+        if (result != null) {
+          if (result!.contains("loginUrl") &&
+              result!.contains("username")) {
+            String jsonData = result!;
             Map<String, dynamic> data = jsonDecode(jsonData);
             final model = QRModel.fromJson(data);
             context.pop({model});
           }
         }
-      }
-    });
-  }
-
-  void _onPermissionSet(BuildContext context, QRViewController ctrl, bool p) {
-    log('${DateTime.now().toIso8601String()}_onPermissionSet $p');
-    if (!p) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text("qr.noPermission".tr())),
-      );
-    }
+      },
+    );
   }
 
   @override
   void dispose() {
-    controller?.dispose();
+    controller.dispose();
     super.dispose();
   }
 }
